Move AppConfig into the shared folder
