+++ server/config.ts
@@ -2,33 +2,10 @@ import toml from '@iarna/toml'
 import debug from 'debug'
 import fs from 'fs/promises'
 
+import { AppConfig } from '../shared/types'
 import { getConfigFilePath } from './paths'
 
 const d = debug('b:config')
-
-// Interface for a single OpenAI provider configuration
-export interface OpenAIProviderConfig {
-  providerName?: string // Name for this provider configuration, the default is 'openai'
-  baseURL?: string
-  apiKey?: string
-  model?: string
-}
-
-// Main application configuration interface
-export interface AppConfig {
-  haBaseUrl?: string
-  haToken?: string
-
-  llm?: string // either 'anthropic', 'ollama', or a provider name in openAIProviders
-
-  anthropicApiKey?: string
-  anthropicModel?: string
-
-  ollamaHost?: string
-  ollamaModel?: string
-
-  openAIProviders?: OpenAIProviderConfig[] // Array for multiple OpenAI configs
-}
 
 export async function createConfigViaEnv(notebookDirectory: string) {
   // Provide a default llm to satisfy the type, migration will fix it
