+++ server/llm.ts
@@ -6,8 +6,8 @@ import { McpServer } from '@modelcontextprotocol/sdk/server/mcp.js'
 import { Observable } from 'rxjs'
 
 import { Automation } from '../shared/types'
+import { AppConfig } from '../shared/types'
 import { AnthropicLargeLanguageProvider } from './anthropic'
-import { AppConfig } from './config'
 import { createHomeAssistantServer } from './mcp/home-assistant'
 import { createMemoryServer } from './mcp/memory'
 import { createNotifyServer } from './mcp/notify'
